Add ChatManager
Move session classes to core. Break out DisputeCommunicationMessage
handling from DisputeManager and put in ChatMananger prepare for other
uses of ChatManager.

Renaming of DisputeCommunicationMessage would be nice but it's
representing the protobuf messages so the name has to stay.
